requirements.txt
appdirs==1.4.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bidict==0.23.1
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
cloud-init==23.3.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.0
coverage==7.4.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
dogtail==0.9.11
entrypoints==0.4
eventlet==0.35.1
exceptiongroup==1.2.0
filelock==3.6.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.3.6
Flask-Testing==0.8.1
fonttools==4.49.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==2.1.0
packaging==23.2
Paste==3.7.1
pexpect==4.8.0
pillow==10.2.0
platformdirs==2.5.1
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==8.0.1
pytest-cov==4.1.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-engineio==4.9.0
python-magic==0.4.24
python-socketio==5.11.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.1.0
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
service-identity==18.1.0
simple-websocket==1.0.0
simplejson==3.17.6
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
threadpoolctl==3.3.0
tomli==2.0.1
Twisted==22.1.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
webapp2==2.5.2
WebOb==1.8.7
Werkzeug==2.3.7
wsproto==1.2.0
xdg==5
zipp==1.0.0
zope.interface==5.4.0
